requirements.txt
regex
pypdf
google-generativeai
python-docx
jupyterlab